% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{captcha_access_trt}
\alias{captcha_access_trt}
\title{O algoritmo funcionaria assim. Primeiro, ele pega qual é a letra que
tem a menor diferença entre o primeiro e o segundo candidatos.
Depois, ele "troca" a posição do primeiro canditado com o segundo.
Se esse teste já foi realizado, ele verifica qual letra seria considerada,
usando o mesmo critério que o anterior (menor diferença entre o primeiro
e o segundo candidatos). Dessa vez, no entanto, ele teria de considerar
o terceiro candidatos.}
\usage{
captcha_access_trt(path)
}
\description{

}
\section{path <- "data-raw/trt/"}{

}

\section{model <- NULL}{
fs::dir_create(path)
u_captcha <- "https://pje-consulta.trt3.jus.br/pje-consulta-api/api/captcha?idProcesso=2104879"
f_captcha <- fs::file_temp(tmp_dir = path, ext = ".jpeg", pattern = "trt")

if (!file.exists(f_captcha)) {
captcha <- httr::GET(u_captcha)
captcha \%>\%
httr::content() \%>\%
purrr::pluck("imagem") \%>\%
base64enc::base64decode() \%>\%
writeBin(f_captcha)
token <- captcha \%>\%
httr::content() \%>\%
purrr::pluck("tokenDesafio")
}

u_consulta <- "https://pje-consulta.trt3.jus.br/pje-consulta-api/api/processos/2104879"
ntry <- 1
label <- captcha_candidates(f_captcha, model, n = max_ntry)
}

\section{browser()}{
f_log <- paste0(dirname(path), "/logs/", fs::path_ext_set(basename(f_captcha), ".log"))
fs::dir_create(dirname(f_log))

acertou <- captcha_trt_test(u_consulta, token, label\link{1})
da_log <- tibble::tibble(
ntry = ntry,
label = label\link{ntry},
type = "auto",
result = acertou
)
if (acertou) {
usethis::ui_done("Acertou!!!")
label <- label\link{ntry}
} else {
max_ntry_model <- min(max_ntry, length(label))
usethis::ui_info("Temos {max_ntry_model} candidatos...")
}

while (!acertou && ntry < max_ntry_model && !is.null(model)) {
usethis::ui_info("Errou! O chute foi: {label\link{ntry}}")
ntry <- ntry + 1
acertou <- captcha_trt_test(u_consulta, token, label\link{ntry})
da_log <- tibble::add_row(
da_log,
ntry = ntry,
label = label\link{ntry},
type = "auto",
result = acertou
)
if (acertou) {
usethis::ui_done("Acertou!!!")
label <- label\link{ntry}
}
}

if (!acertou && !manual) {
label <- label\link{ntry}
}
}

\section{if tried {max_ntry} times and the model still did not find it}{

}

