% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sei.R
\name{captcha_download_sei}
\alias{captcha_download_sei}
\title{Download SEI}
\usage{
captcha_download_sei(path, classify = FALSE)
}
\arguments{
\item{path}{path to folder where file will be saved}

\item{classify}{optional. Automatically classify captcha.}
}
\description{
Download SEI
}
