% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-oracle.R
\name{captcha_oracle}
\alias{captcha_oracle}
\title{Função que baixa um captcha e classifica usando oráculo}
\usage{
captcha_oracle(
  path,
  model = NULL,
  max_ntry = 10,
  manual = TRUE,
  captcha_access,
  captcha_test
)
}
\arguments{
\item{path}{caminho em que os arquivos serão salvos}

\item{model}{modelo para predizer a label de uma imagem}

\item{max_ntry}{quantidade máxima de chutes}

\item{manual}{caso o máximo de tentativas seja alcançado, abrir o prompt
para classificar manualmente?}

\item{captcha_access}{função que baixa um captcha e retorna dados da
sessão para validar o captcha}

\item{captcha_test}{função que testa se um captcha está correto
a partir de uma label específica}
}
\description{
Essa função é muito relevante para construção dos resultados, pois
é responsável por fazer as classificações usando o oráculo dos sites
}
