% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{captcha_oracle_trt_com_feedback}
\alias{captcha_oracle_trt_com_feedback}
\title{Classifica captchas com feedback do oraculo}
\usage{
captcha_oracle_trt_com_feedback(
  path,
  model = NULL,
  max_ntry = 10,
  manual = TRUE
)
}
\arguments{
\item{path}{path}

\item{model}{model}

\item{max_ntry}{max tries}

\item{manual}{whether to annotate manually when model fails}
}
\description{
Classifica captchas com feedback do oraculo
}
