% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{captcha_candidates}
\alias{captcha_candidates}
\title{O algoritmo funcionaria assim. Primeiro, ele pega qual é a letra que
tem a menor diferença entre o primeiro e o segundo candidatos.
Depois, ele "troca" a posição do primeiro canditado com o segundo.
Se esse teste já foi realizado, ele verifica qual letra seria considerada,
usando o mesmo critério que o anterior (menor diferença entre o primeiro
e o segundo candidatos). Dessa vez, no entanto, ele teria de considerar
o terceiro candidatos.}
\usage{
captcha_candidates(f_captcha, model, cut_value = log(0.01), n)
}
\arguments{
\item{f_captcha}{captcha file}

\item{model}{model that generates logits}

\item{cut_value}{so that we make combinations from real values}

\item{n}{maximum number of candidates}
}
\description{
Vamos pensar em outra heurística: menor probabilidade.
Nessa heurística, selecionamos a letra com menor probabilidade
e pegamos o segundo candidato dessa letra.
Calculates n captcha candidate labels
}
\details{
Using all captcha combinations from some cut value
}
